% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Gauss}
\alias{Gauss}
\title{Generates a Gauss gAlg function}
\usage{
Gauss(sd = 1, mean = 0)
}
\arguments{
\item{sd}{Vector of standard deviations (for each dimension)}

\item{mean}{Vector of means (for each dimension)}
}
\description{
Creates a gAlg function representing: f(x1,x2,...xd) = Normal distribution(sd,mean)
}

