% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{gaussAlgebra-package}
\alias{gaussAlgebra-package}
\title{Package for efficient analitical calculations on polynomial-times-gaussian functions}
\description{
Package for efficient analitical calculations on polynomial-times-gaussian functions
}

