% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linear.gAlg}
\alias{linear.gAlg}
\title{Generates a linear gAlg function}
\usage{
linear.gAlg(k, d = 1)
}
\arguments{
\item{k}{Index of the dimension in which the function is linear}

\item{d}{Dimension in which the function lives}
}
\description{
Creates a gAlg function representing: f(x1,x2,...xd) = xk
}

