% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussAlgebra.R
\name{poly}
\alias{poly}
\title{Transforms polynomial coefficients into a 1D gAlg function}
\usage{
poly(...)
}
\arguments{
\item{...}{coefficients of a polynomial}
}
\description{
Transforms polynomial coefficients into a 1D gAlg function
}

