% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussAlgebra.R
\name{print.gAlg}
\alias{print.gAlg}
\title{Prints a gAlg function in a nice readable form}
\usage{
\method{print}{gAlg}(x, ...)
}
\arguments{
\item{x}{the gAlg function to be printed}
}
\description{
Prints a gAlg function in a nice readable form
}

