% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{calc,gAlg,numeric-method}
\alias{calc,gAlg,numeric-method}
\title{Evaluates the gAlg function in specific points}
\usage{
\S4method{calc}{gAlg,numeric}(object, x)
}
\arguments{
\item{object}{gAlg function}

\item{x}{points in which to evaluate the function}
}
\description{
Evaluates the gAlg function in specific points
}

