% Generated by roxygen2 (4.0.2): do not edit by hand
\name{const.gAlg}
\alias{const.gAlg}
\title{Generates a constant gAlg function}
\usage{
const.gAlg(a = 1, d = 1)
}
\arguments{
\item{a}{Value of the constant function}

\item{d}{Dimension in which the function lives}
}
\description{
Creates a gAlg function representing: f(x1,x2,...xd) = a
}

