% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gAlg-class}
\alias{gAlg}
\alias{gAlg-class}
\title{Class that represents a polynomial-times-gaussian function}
\description{
Class that represents a polynomial-times-gaussian function
}
\section{Slots}{

\describe{
\item{\code{tab}}{Array keeping all the information about the function (see below)}
}}

