% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gaussAlgebra.R
\docType{package}
\name{gaussAlgebra-package}
\alias{gaussAlgebra-package}
\title{Package for efficient analitical calculations on polynomial-times-gaussian functions}
\description{
Package for efficient analitical calculations on polynomial-times-gaussian functions
}

